requirements/base.txt
Django==3.1.5
django-oauth-toolkit==1.3.3
django-redis==4.12.1
djangorestframework==3.12.2
ramos==1.4.0
zeep==4.0.0